pyproject.toml
[tool.poetry]
name = "gym"
version = "0.1.0"
description = ""
authors = ["Дмитрий <55105946+isys35@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
requests = "^2.27.1"
psycopg2 = "^2.9.3"
commitizen = "^2.25.0"
python-dotenv = "^0.20.0"
bs4 = "^0.0.1"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.commitizen]
name = "cz_customize"

[tool.commitizen.customize]
message_template = "{{change_type}}:{% if show_message %} {{message}}{% endif %}"
example = "👨‍💻 FEAT: Изменяем вид коммитов"

[[tool.commitizen.customize.questions]]
type = "list"
name = "change_type"
choices = ["👨‍💻 FEAT", "🐞 FIX", "🛠 REFACTOR", "🔨 TEST", "📚 DOCS"]
message = "Select the type of change you are committing"

[[tool.commitizen.customize.questions]]
type = "input"
name = "message"
message = "Тело сообщения коммита"

[[tool.commitizen.customize.questions]]
type = "confirm"
name = "show_message"
message = "Вы хотите добавить сообщение коммита?"